allow for combos of gitbook save
